refactor: delete BackgroundCircle, and change previous components using it
