+++ src/components/@common/BackgroundWithCircle/BackgroundWithCircle.tsx
@@ -1,7 +1,19 @@
 import type { SxProps } from '@mui/material';
 import { Box } from '@mui/material';
-import BackgroundCircle from 'components/background-circle';
 import type { PropsWithChildren } from 'react';
+
+const BackgroundCircle = () => (
+  <Box
+    width='35.9375rem'
+    height='36.6875rem'
+    bgcolor='p.8'
+    borderRadius='50%'
+    zIndex='2'
+    sx={{
+      filter: 'blur(25px)',
+    }}
+  ></Box>
+);
 
 export const BackgroundWithCircle = ({ children, sx }: PropsWithChildren<{ sx: SxProps }>) => (
   <Box width={'100vw'} position={'relative'} bgcolor={'#f9f9f9'} sx={sx}>
