--- src/components/background-circle/index.ts
@@ -1,3 +1,0 @@
-import BackgroundCircle from './BackgroundCircle';
-
-export default BackgroundCircle;
